add text editor for Javascript
The editor will be hidden unless showTextEditor=1 is passed to the page.

The Javascript will be evaluated in the context of the main
page so evil Javascript could do all kinds of mean shenanigans.
Therefore, we don't run it unless the user presses ctrl+enter in
the text box.

We should add some clear warnings about the risks of running untrusted
Javascript and we will need a way to disable it for server
operators if this feature should ever have a chance to get into the
upstream code base. It should never be enabled on any domain that hosts
other services (because Javascript can steal cookies and maybe even mess
with other pages via web worker pre-caching).