add button for running Javascript
The keyboard shortcut is more useful in practice but the button is
more obvious for new users (who may have failed to read the message
that mentions the shortcut). We may also want to use this on mobile
devices (at least to run existing code) and the button is much easier
on a touch-only device.